clean up header component and add edit button
