+++ packages/slice-machine/lib/builders/CustomTypeBuilder/Layout/index.tsx
@@ -1,3 +1,3 @@
-export { default as Header } from "./Header";
+export { default as Header } from "./Header/index";
 export { default as Tabs } from "./Tabs";
 export { default as FlexWrapper } from "./FlexWrapper";
